revamped the requests modals
